Ensemble vm working with different seeds for each ensemble member and corrected model saving process. Fix_val validation is excluded as it is intended in a different pipeline.

+++ training/.ipynb_checkpoints/main-checkpoint.py
@@ -1,4 +1,4 @@
-import argparse
+
 import getpass
 import os
 import shutil
@@ -10,58 +10,41 @@ from contextlib import contextmanager
 
 import wandb
 
-from training.runner import Runner
-from training.utilities import GeneralUtility
-
-# Set logging levels for external libraries
-import logging
-logging.getLogger('matplotlib').setLevel(logging.WARNING)
-logging.getLogger('PIL').setLevel(logging.WARNING)
-logging.getLogger('urllib3').setLevel(logging.WARNING)
-logging.getLogger('wandb').setLevel(logging.WARNING)
+from runner import Runner
+from utilities import GeneralUtility
 
 warnings.filterwarnings('ignore')
 
-# Parse command-line arguments
-parser = argparse.ArgumentParser(description="Training script for canopy height prediction.")
-parser.add_argument("--debug", action="store_true", help="Run in debug mode.")
-parser.add_argument("--model_save_dir", type=str, default="./models", help="Directory to save trained models.")
-args = parser.parse_args()
-
-debug = args.debug
-model_save_dir = args.model_save_dir
-
-# Ensure the model save directory exists
-os.makedirs(model_save_dir, exist_ok=True)
-
+debug = "--debug" in sys.argv
 
 defaults = dict(
     # Model save directory
-    model_save_dir=model_save_dir,
+    model_save_dir="./models",
 
     # System
-    seed=1,
+    seed=1, # bei ensemble rausnehmen!!
 
     # Data
-    dataset='/dataset',#'ai4forest_debug',
-    batch_size=5,
-
-    # Model variant specifics
-    ensemble_size=3,
-    loss_name='l2',#'shift_huber',  # Defaults to shift_l1
+    dataset='', #ai4forest_debug
+    batch_size=16,
 
     # Architecture
     arch='unet',  # Defaults to unet
     backbone='resnet50',  # Defaults to resnet50
     use_pretrained_model=False,
 
+    # Model variant specifics
+    variant='baseline', # 'ensemble', 'baseline'
+    ensemble_size=3,
+    loss_name='l2',  # Defaults to shift_l1
+
     # Optimization
     optim='AdamW',  # Defaults to AdamW
-    n_iterations=5, #100
-    log_freq=5,
+    n_iterations=50000, #100, 25000
+    log_freq=2000,
     initial_lr=1e-3,
     weight_decay=1e-2,
-    use_standardization=False,
+    use_standardization=False, #default: False
     use_augmentation=False,
     use_label_rescaling=False,
 
@@ -81,22 +64,30 @@ defaults = dict(
     cyclic_mode='triangular2',
     )
 
+print(f"Defaults: {defaults}")
+
+"""
 if not debug:
     # Set everything to None recursively
     defaults = GeneralUtility.fill_dict_with_none(defaults)
+"""
 
 # Add the hostname to the defaults
 defaults['computer'] = socket.gethostname()
+print(f"Defaults: {defaults}")
+
+
 
 # Configure wandb logging
 wandb.init(
     config=defaults,
-    project='test-000',  # automatically changed in sweep
+    project='base-001',  # automatically changed in sweep
     entity=None,  # automatically changed in sweep
 )
 config = wandb.config
 config = GeneralUtility.update_config_with_default(config, defaults)
 
+print(f"WandB Config: {config.items}")
 
 @contextmanager
 def tempdir():
@@ -133,15 +124,23 @@ with tempdir() as tmp_dir:
     runner.run()
 
     # Save the trained ensemble models
-    for idx, model_path in enumerate(runner.model_paths['ensemble']):
-        permanent_path = os.path.join(config.get('model_save_dir', './models'), f'ensemble_model_{idx}.pt')
+    if config.variant == 'ensemble':
+        for idx, model_path in enumerate(runner.model_paths['ensemble']):
+            permanent_path = os.path.join(config.get('model_save_dir', './models'), f'test_ensemble_model_{idx}.pt')
+            shutil.copy(model_path, permanent_path)
+            print(f"Saved ensemble model {idx+1} to {permanent_path}")
+    
+    else:
+        variant = config.variant
+        model_path = runner.model_paths[f'{variant}']
+        permanent_path = os.path.join(config.get('model_save_dir', './models'), f'{variant}_model.pt')
         shutil.copy(model_path, permanent_path)
-        print(f"Saved ensemble model {idx+1} to {permanent_path}")
+        print(f"Saved {variant} model to {permanent_path}")
 
-    # Close wandb run
-    wandb_dir_path = wandb.run.dir
-    wandb.join()
-
-    # Delete the local files
-    if os.path.exists(wandb_dir_path):
-        shutil.rmtree(wandb_dir_path)
+        # Close wandb run
+        wandb_dir_path = wandb.run.dir
+        wandb.join()
+    
+        # Delete the local files
+        if os.path.exists(wandb_dir_path):
+            shutil.rmtree(wandb_dir_path)
